EmailAndPasswordInputPageクラスを追加
